WIP: template builder
