--- De.Loooping.Templates.Core/Configuration/IContentReplacerConfiguration.cs
@@ -1,8 +1,0 @@
-namespace De.Loooping.Templates.Core.Configuration;
-
-public interface IContentReplacerConfiguration
-{
-    bool AllowFormatting { get; }
-    bool AllowChildren { get; }
-    char ChildSeparator { get; }
-}
