+++ De.Loooping.Templates.Core/Template/CompilerErrorException.cs
