+++ De.Loooping.Templates.Core/Template/SyntaxException.cs
